Flatmate's Bill/flatmate_bill/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
filestack-python==3.5.0
fpdf==1.7.2
idna==3.6
requests==2.31.0
trafaret==2.0.2
urllib3==2.1.0
